FIX: fixed REPL so that it works with svelvet6
There were two issues:
(1) REPL is incompatible with uuid. This is potentially a serious issue; we are risking pk collisions so that the REPL will work
(2) For reasons unknown, the REPL does not like it when minimap uses "findStoreById" to access the store. It requires that the store be accessed through props or directly

For the time being, I have made changes so that Svelvet6 is compatible with the REPL. However, the next team should think about whether the technical debt of the REPL is worth risking pk collisions