FIX: input fields can now be interacted with when embedded into Node
Node previously disabled default behavior on mouse down. This was to prevent dragging images embedded in node. However, this made input fields impossible to click on when embedded into node. To fix this, we disable pointer events on images embedded in nodes, and allow default behavior when clicking within a node
With this fix, we can respond to: https://old.reddit.com/user/Mopetus